Fix a bug that caused uncompilable code to be generated
The bug could cause an invalid provider to be picked if there was a
provider that provides a subclass with fewer generic type parameters
existed.

So for example if you had a property of the type `Map<String, String>`
and a subclass to map called `IntMap<T> : Map<Int, T>` and a generic
provider for that subclass then that provider would be used and thus the
generated code would not compile.