Improve how to implement the factory
Instead of having to make the factory abstract and have a static
INSTANCE field factories can now be an `object` that delegates the
adapter creation to the generated factory.

This was done previously but there was issues but those seems to have
been caused by SuperficialValidation.validateElement returning false for
the factory class in some cases. It doesn't seem like that call is
needed so now the generation has been simplified.